src/django_backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
Pillow==8.0.0
promise==2.3
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.1
psycopg2-binary==2.8.6